// apps/profile/src/github-scrapper/dtos/scrap-github-profile.dto.ts
export class ResponseScrapGithubProfileDto {
  avatar_url: string;
  html_url: string;
  repos_url: string;
  name: string;
  company: string;
  blog: string;
  location: string;
  email: string;
  bio: string;
  public_repos: number;
  followers: number;
  following: number;
  created_at: string;
}